nodejs: converting dialog window

+++ web/documentserver-example/nodejs/app.js
@@ -319,7 +319,7 @@ app.post('/convert', (req, res) => { // define a handler for converting files
   const lang = req.body.lang ? req.body.lang : null;
   const fileUri = req.DocManager.getDownloadUrl(fileName, true);
   const fileExt = fileUtility.getFileExtension(fileName, true);
-  const internalFileExt = 'ooxml';
+  const internalFileExt = req.body.fileExt ? req.body.fileExt : 'ooxml';
   const response = res;
 
   const writeResult = function writeResult(filename, step, error) {
@@ -367,20 +367,31 @@ app.post('/convert', (req, res) => { // define a handler for converting files
       if (status !== 200) throw new Error(`Conversion service returned status: ${status}`);
 
       // write a file with a new extension, but with the content from the origin file
-      fileSystem.writeFileSync(req.DocManager.storagePath(correctName), data);
-      fileSystem.unlinkSync(req.DocManager.storagePath(fileName)); // remove file with the origin extension
+      if (fileUtility.getFileType(correctName) !== "other") {
+        fileSystem.writeFileSync(req.DocManager.storagePath(correctName), data);
+      } else {
+        writeResult(newFileUri, result, "FileTypeIsNotSupported");
+        return;
+      }
+      if (!("fileExt" in req.body)) fileSystem.unlinkSync(req.DocManager.storagePath(fileName)); // remove file with the origin extension
 
       const userAddress = req.DocManager.curUserHostAddress();
       const historyPath = req.DocManager.historyPath(fileName, userAddress, true);
       // get the history path to the file with a new extension
       const correctHistoryPath = req.DocManager.historyPath(correctName, userAddress, true);
 
-      fileSystem.renameSync(historyPath, correctHistoryPath); // change the previous history path
+      if(!("fileExt" in req.body)){
+        fileSystem.renameSync(historyPath, correctHistoryPath); // change the previous history path
 
-      fileSystem.renameSync(
-        path.join(correctHistoryPath, `${fileName}.txt`),
-        path.join(correctHistoryPath, `${correctName}.txt`),
-      ); // change the name of the .txt file with document information
+        fileSystem.renameSync(
+          path.join(correctHistoryPath, `${fileName}.txt`),
+          path.join(correctHistoryPath, `${correctName}.txt`),
+        ); // change the name of the .txt file with document information
+      } else if (newFileType !== "other") {
+        const user = users.getUser(req.query.userid);
+
+        req.DocManager.saveFileData(correctName, user.id, user.name);
+      }
 
       writeResult(correctName, result, null); // write a file with a new name to the result object
     } catch (e) {
@@ -391,7 +402,7 @@ app.post('/convert', (req, res) => { // define a handler for converting files
 
   try {
     // check if the file with such an extension can be converted
-    if (fileUtility.getConvertExtensions().indexOf(fileExt) !== -1) {
+    if ((fileUtility.getConvertExtensions().indexOf(fileExt) !== -1) || ("fileExt" in req.body)) {
       const storagePath = req.DocManager.storagePath(fileName);
       const stat = fileSystem.statSync(storagePath);
       let key = fileUri + stat.mtime.getTime();
